backend/requirements.txt
celery==5.3.6
Django==4.2.10
django-ckeditor==6.7.0
django-debug-toolbar==4.3.0
easy-thumbnails==2.8.5
Pillow==8.3.2
pytz==2018.4
redis==5.0.1
singlemodeladmin==0.8
vine==5.1.0
gunicorn==21.0.1
django-modeltranslation==0.18.11
crispy-bootstrap5==2024.10
django-crispy-forms==2.3
